hams_notification/requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
Django==4.2
django-environ==0.10.0
frozenlist==1.3.3
idna==3.4
kombu==5.2.4
multidict==6.0.4
pika==1.3.1
prompt-toolkit==3.0.38
PyJWT==2.6.0
python-dotenv==1.0.0
pytz==2023.3
requests==2.29.0
six==1.16.0
sqlparse==0.4.4
twilio==8.1.0
tzdata==2023.3
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
yarl==1.9.2
